Upload packaged app on non PR runs
Whenever a new master commit, scheduled or triggered build run is executed, we now upload the packaged app to the artifacts.
We already did that for the mac build, because we need it for the publishing, but now we do it for the other systems as well.
But only on non PRs to not increase the PR build time.

Signed-off-by: Simon Graband <sgraband@eclipsesource.com>